Merged PR 277418: Added ResourcesDBClientProvider and configuration repo in resources db
Fixes: https://github.com/microsoft/vssaas-planning/issues/858

Description of the Change
-------------------------

[Provide as many details as possible about what the code is doing, think about folks unfamiliar with the codebase]

Added ResourcesDBClientProvider - Regional and the Global, and configuration repo in resources db.

Impact
------

No impact is expected. The change only creates a database in the Resources database account and then instantiates the creation of "configuration" collection inside the database.

Rollback Plan
-------------

[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]

The changes can be rolled back by just reverting the commit.