Merged PR 278114: Enable VsoUtil commands to query VsVersionUrl and VsVersion from SkuCatalog
Description of the Change
-------------------------

This enables a couple of new commands in VSoUtil that grabs the VsChannelUrl and VsVersion properties for the Sku Catalog, so they can be used as part of Nexus image promotion pipline.

Impact
------

VsoUtil tool change enabling new functionality, so little impact.  Tool is used in build pipelines.

Rollback Plan
-------------

Just revert the commit.