Merged PR 278132: Add GitHubPortForwardingServiceEnabled secret to main.tf
Fixes: found while setup up a local environment

Description of the Change
-------------------------

This is setting a "local" secret used by main.tf which allows terraforming of personal dev cluster.
This is required by the `portal-vsclk-portal-website` pod to run.
A secret was likely added without updating the main.tf file.

Impact
------

This fix isn't only for local dev, it has no impact on prod.

Rollback Plan
-------------

Commit can simply be reverted.