Merged PR 278160: chore: Understanding user errors 1151 and 1152
Fixes: https://github.com/microsoft/vssaas-planning/issues/782

Description of the Change
-------------------------

chore: Understanding user errors 1151 and 1152

The errors 1151 and 1152 happens because IGitCredentialService is not available at the time of cloning and building private docker containers etc. The reason why it wasn't available was, the Client was not available at that time. Plotting the values, we see a pattern, where users wait for 8 seconds and close the client (browser).
```
ResumeOutcomeClassifiedEvents
| where TimeStamp > ago(30d)
| where ServiceEnvironment == "vsclk-online-prod"
| where ResumeOutcome contains "Infrastructure: CODE:115"
| project Duration = (EventAtStartingToShutdown_Time - EventAtResumeRequestStart_Time)
| summarize count() by bin(Duration, 1s)
| render columnchart
```

Impact
------

It only updates kusto queries. No change to code.

Rollback Plan
-------------

DRI can simple revert this change.