Merged PR 277882: Port heartbeat continuationn handlers to new Job Queue API
Fixes: [link to issue]

Migrate existing continuation handlers to use the new Job queue API by creating new IJobTargetHandler instances that will be started during the front end startup.
The implementation was simplified by splitting the creation of the payload and the job handler implementation. Please note this new code will be activated when this condition is evaluated ->
        private Task<bool> IsJobContinuationHandlerEnabledAsync(IDiagnosticsLogger logger)
        {
            return ConfigurationReader.ReadFeatureFlagAsync("job-continuation-handler", logger, false);
        }

Impact
------
Heartbeat monitoring. It is needed as part of the scalability migration.

[How critical is the fix, think potential hotfix justification]

Rollback Plan
-------------
The default for all stamps is disabled unless we enable this on the front end database.
feature:vsclk-usw2:job-continuation-handler-enabled

[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]