Merged PR 278346: Adding shallow clone feature flag
Fixes: https://github.com/microsoft/vssaas-planning/issues/872

Description of the Change
-------------------------

This change adds a feature flag that can be passed to the agent to specify whether or not the agent should attempt a shallow clone before executing `git fetch --unshallow` in the background to optimize Codespace creation time. See corresponding Cascade PR: https://dev.azure.com/devdiv/DevDiv/_git/Cascade/pullrequest/277824.

Impact
------

Low; this change simply adds a feature flag that is false by default. Toggling the feature flag to be true may have higher impact as described in the PR above.

Rollback Plan
-------------

Revert this commit AND the commit in the above PR.