Merged PR 278648: Add flag to enable deleting orphaned system resources by resource type
Fixes: [link to issue]

Description of the Change
-------------------------

Currently `ConfirmOrphanedSystemResourceJobHandler` deletes all resources it finds to be orphaned which makes it hard to verify correctness without risking resources. This adds a feature flag to the handler to enable deletion on a per-resource-type basis so that we can enable the task for some resources while validating it for others.

Impact
------

Hotfix not needed

Rollback Plan
-------------

If this commit is reverted then we need to validate that the flag in the BE DB to enable `WatchOrphanedSystemResourceJobHandler` is *disabled* as after this change rolls out it will probably be enabled. If this is rolled back when that flag is enabled we would begin deleting all resource types.