Merged PR 278646: Even in phase 2, fall back to GlobalRepository.GetAsync()
Fixes: This is a fixup to https://devdiv.visualstudio.com/OnlineServices/_git/vsclk-core/pullrequest/275671

Description of the Change
-------------------------

When setting the regional cloud environment db rollout-status configuration property to "2", actions like Resume (and presumably Suspend) would fail for codespace environments that lived in a regional location outside of the one that the client is connected to.

This would happen because the RegionalRepository.GetAsync() query would return null which would cause the EnvironmentGetAction code to throw a ItemNotFound exception (which corresponds to an HTTP 404).

What this fix does is to fall back to getting the GlobalRepository record for the codespace environment if it does not exist in the regional db so that we can properly issue a RedirectToLocationException so that the client is properly redirected to the appropriate region.

Impact
------

This does not need to be a hotfix because none of the PROD services are configured to use the phase "2" code path (and actually, the PR that this fixes isn't even rolled out yet anyway).

Rollback Plan
-------------

This commit itself doesn't change the behavior of "Phase 1" which is what is currently being used. This only changes the behavior of "Phase 2" which is not yet in use.

The way to change the rollout phase is to edit the "configuration" records in the global cosmosdb for the following:

```
{
   "id": "setting:vsclk-{region-code}:cloudenvironmentrepository-rollout-status",
   "value": ["1", "2", ...]
}
```

By setting the value to "1", it will roll back to proven code. Setting the value to "2" will run the newer logic.

Even in phase 2, fall back to GlobalRepository.GetAsync()

Fixes Connect to Codespace/Suspend/Resume.