Merged PR 278531: Integrate the new vssdk package and use it on the Azure Orphan resource job
Fixes: [link to issue]

Fix the telemetry on the Azure Orphan delete job handlers.
Please notice a new diagnostic library is being used from the vssdk that use a different implementation on how to maintain the LogValueSet.

Impact
------

[How critical is the fix, think potential hotfix justification]

Rollback Plan
-------------
Revert the commit is you find anything suspicious on some unrelated side effect of vssdk upgrade.

Revert the feature flag in case it was enabled.

Integrate the new vssdk package and use it on the Azure Orphan resource job handler