Merged PR 279222: [PF] Fix cookie invalidation
Fixes: found while working on similar code for PF error backend

Description of the Change
-------------------------

Invalidates the __Host-vso-pf cookie properly by adding passing HttpOnly and Secure. These attributes are required by any __Host cookie to be changed.

Impact
------

This didn't come up in an issue but it might cause issues to customers as this looks to be a regression.

Rollback Plan
-------------

Yes commit can simply be reverted.