Merged PR 279197: Fix continuation job expire time out
Fixes: [link to issue]

The way we track today job continuation expiration is by checking the payload job creation but NOT the original created continuation payload. That mean if a job continuation is retrying on a specific state it will re queue the payload on a retry ad so resetting the expire time out. Bottom line, it can run forever without any limits and we want to fix this ASAP.

Impact
------

[How critical is the fix, think potential hotfix justification]

Rollback Plan
-------------

Most of this code i still under a feature flag and so there should be no need to revert this commit.