Merged PR 279492: Add check for missing OSDisk Component.
Fixes: [link to issue]

Description of the Change
-------------------------
Adding check to verify if OSDisk component is missing in windows VM, in heartbeat handler:
 -  dont mark it ready or assign the faulty VM
 -  delete faulty VM
[Provide as many details as possible about what the code is doing, think about folks unfamiliar with the codebase]

Impact
------
This will clean up the faulty windows VMs
[How critical is the fix, think potential hotfix justification]

Rollback Plan
-------------
yes, it can be reverted. It is for a bug fix, so when reverted, we will see the issue that windows compute is assigned to codespace where OSDisk component is missing. It will cause codespaces to stay stuck in Queued state.
[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]

Add check for missing OSDisk Component.