Merged PR 279673: Disable GH API forking for Nexus
Fixes: [link to issue]

Description of the Change
-------------------------

Fixes a release blocking issue where Cascade token user ids mismatch in Nexus when a codespace is created through the GH gateway. Disables all codepaths which use the gateway.

Impact
------

Release blocker fix

Rollback Plan
-------------

Revert it or follow the comments to delete the changes manually