Merged PR 280062: [Web] Fix for CORS failures in serverless
Fixes: [link to issue]

Description of the Change
-------------------------

Fix for CORS failures in serverless
The CORS failures started happening as web worker extension host is wrapped in iframe now due to security reasons.
https://github.com/microsoft/vscode-internalbacklog/blob/master/serverless-release-notes/v1_50.md#web-worker-extension-host-wrapped-in-iframe
This is a temporary change to unblock ourselves having LiveShare/Serverless working. The more permanent fix would be to move to CDN (and another temporary change could be to return access-control-allow-origin with * only for these assets)

Impact
------

Serverless scenarios are currently broken including LiveShare web

Rollback Plan
-------------
Revert the commit