Merged PR 279207: New Port Forwarding Error Backend
Fixes: https://github.com/microsoft/vssaas-planning/issues/202

Description of the Change
-------------------------

Add a new service to our cluster: error backend.
Nginx is now configured to re-route some errors to our custom error backend so we can serve pretty HTML error pages. Those include short descriptions of the error, links to the documentation and service status page.
In addition, the error backend is sending a response that clears the Port Forwarding auth cookie (which can be responsible of some errors).

Impact
------

This is a large change (since it includes the deployment of a new kubernetes service) but it's mostly cosmetic.

Rollback Plan
-------------

This commit pushes a new service to the cluster named `web-errors-backend` and pod named `web-errors-backend-[hash]`. Pipelines needs to be updated to deploy such change, they will need to be update to roll it back too. If this needs to be reverted talk to @<Peter Lisy> or @<Vincent Dondain>.

Due to the nature of the change, this could very well be rolled back if needed.