Merged PR 280092: Log all states in LogCloudEnvironmentStateTask
Description of the Change
-------------------------

`LogCloudEnvironmentStateTask` queries the counts of environments in all states, but only persists the count for a few of them. This change improves the telemetry so counts of environments in states like `Failed` and `Unavailable` are also logged. We are accumulating a number of environments in transitional phases as well, such as `Starting`, and it would be helpful to be able to see this in the telemetry as well. Because the Cosmos query already grabs all the data, the query itself would be no more expensive.

Impact
------

Low. Improves the telemetry.

Rollback Plan
-------------

Revert commit.