Merged PR 278720: Add anonymous token support to Token Service
Description of the Change
-------------------------

LiveShare needs to issue anonymous tokens for the use case of guests connecting to a session without having a Live Share profile. This connection is a read-only one.
That token is currently created by the Node js auth service, and we want to migrate that logic to our Token Service.

Impact
------

Having the Token Service issuing anonymous tokens requires to configure the Issuer and Audience for such tokens.
Also, we need to make some changes in the consumers of this endpoint, including using the jwt token ID `jti` in the same way it's currently used a UserId claim, included in the auth svc anonymous tokens, but not in our Token service issued ones.

Rollback Plan
-------------

We can simply revert this commit, and any changes done on the consumers if any.