Merged PR 280152: Update splash screen package version
Fixes: [link to issue]

Description of the Change
-------------------------

This just updates the splash screen package version to add support for copying text out of the terminals in the creation log.

Impact
------

This enhances the user experience.

Rollback Plan
-------------

Do roll back would be simply going back to the previous version of the splash screen.