Merged PR 279432: A few fixes for the VS updates feature
Fixes: https://github.com/microsoft/vssaas-planning/issues/880

Description of the Change
-------------------------

This fixes a few small issues found in telemetry after a few days of the VS updates feature
being enabled in dev.

Impact
------

Feature is still being tested and only enabled on dev, so shouldn't need any hot fix for now.

Rollback Plan
-------------

If anything wrong is found with the VS updates feature, and since it is hidden behind a feature
flag, just disabling the whole feature should be enough.