Merged PR 280469: Update to splash screen package
Fixes: [link to issue]

Description of the Change
-------------------------

Just an update for the splash screen. This fixes an issue with copying support for xterm.

Impact
------

Without it, copying out of the terminals on the creation log will not work.

Rollback Plan
-------------

We can simply revert the change and consume an older package of the splash screen.