Merged PR 280849: Split hitcount logging into chunks so that kusto can log
Fixes: [link to issue] https://github.com/microsoft/vssaas-planning/issues/1034

Description of the Change
-------------------------

[Provide as many details as possible about what the code is doing, think about folks unfamiliar with the codebase]

The hit count is not being logged in the prod environment because it is huge in size. This code will split the logging into multiple entries so that we could easily log them.

Impact
------

[How critical is the fix, think potential hotfix justification]
No impact is expected. This change only deals with log

Rollback Plan
-------------

[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]

Can be reverted easily without any issues. Only one lens dashboard will get affected.