Merged PR 281156: Fix serializer on the new creation resource payload
Fixes: [link to issue]

Recent testing on the dev ci stamp reveal some issues when using some type of payloads on the creation resource code path. The root cause was that the new Job Continuation API prohibit the use of 'auto type handling' and so some properties were not able to be properly deserialized form the azure queue causing 'invalid cast exception' errors.

The solution is to increment the scope of the special json converters that helps with the serialization and also allow a 'hybrid' model were if a new type is missing will still use the 'auto type handling'.
Impact
------

This will unblock dev to be tested.

Rollback Plan
-------------

[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]