Merged PR 281191: Billing V2: Fixing Feature Flag Check
Description of the Change
-------------------------

When the feature flag checks were moved in Billing V2, the archiving flag was missed. This is impacting performance as the number of summaries being churned through in the scrubber is rising steadily:

![Capture.PNG](https://dev.azure.com/devdiv/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/281191/attachments/Capture.PNG)

Additionally, because of our low default page size in Cosmos, this is translating into a large number of requests (retrieving 250 summaries would take 3 requests, for example).

![Capture2.PNG](https://dev.azure.com/devdiv/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/281191/attachments/Capture2.PNG)

Impact
------

Medium, improves performance by archiving

Rollback Plan
-------------

Turn off feature flag, or rollback commit.