Merged PR 281219: [Codespaces] Only install essential extensions by default
Description of the Change
-------------------------

This PR updates the Codespaces initialization process, so that we only automatically install extensions that are truly essential. Once we do the work to make Live Share feel more native/integrated into GitHub, then we'll add it back to the default list. For now, this change will allow us to deeply focus on the core Codespaces experience as we drive towards GA.

Rollback Plan
-------------

Simply revert this change. I left the list of non-essential extensions in tact, so that when we're ready, we can add them back into the first-run setup experience.