Merged PR 281466: chore: reliability potential fix by disabling transparent huge pages
Fixes: [link to issue]

Description of the Change
-------------------------

chore: reliability potential fix for disabling trasparent huge pages

Based on literature on the internet, THP is an optimization but also is plagued with bugs. Potential side effect being OOM killer killing off process which are important for us.
This change turns off THP.  At best THP would increase the perf by 10% but it has the added risk of unpredictability kill apps. We can consider turning this on, if it didn't improve the reliability.

Impact
------

10% of the resumes are failing. Some of it is attributed to, processes being killed - potentially because of OOM. Since it is hard to investigate this after the fact, this is a potential fix - if it works we will keep it. If it doesn't we will revert it.

Rollback Plan
-------------
This is a potential fix - if it works we will keep it. If it doesn't we will revert it.