Merged PR 281203: Changing the CurrentImageInfoProvider to use ConfigurationReader
Fixes: [link to issue] https://github.com/microsoft/vssaas-planning/issues/1036

Description of the Change
-------------------------

[Provide as many details as possible about what the code is doing, think about folks unfamiliar with the codebase]

This changes CurrentImageInfoProvider to make use of ConfigurationReader API instead of directly making use of SystemConfiguration. This will change the keys that the code uses for lookup.

example of keys -
1. **setting:vsclk:images-vmagent-vsoagentlinux-name** (global scope, least priority), **setting:vsclk-usw2:images-vmagent-vsoagentlinux-name** (region scope, highest priority)
2. **setting:vsclk:images-compute-nexusInternalWindowsImage-name** (global scope, least priority), **setting:vsclk-usw2:images-compute-nexusInternalWindowsImage-name** (region scope, highest priority)
Impact
------

[How critical is the fix, think potential hotfix justification]
This is critical for migration to global BE database for configurations. The [PR 280867](https://devdiv.visualstudio.com/OnlineServices/_git/vsclk-core/pullrequest/280867) writes to the Global BE DB in only the format that could be understood by ConfigurationReader. If we move to Global DB and this change is not present, the code would make queries to global database with keys in old format which won't be present there and hence we would always return the default values.

Rollback Plan
-------------

[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]
Reverting this change means the code is going to make use of old format of  keys which won't be present in global database. So the DRIs would be required to do one of the following two to make sure they get the desired result:
1. Revert back and start using regional configuration database - https://devdiv.visualstudio.com/OnlineServices/_git/vsclk-core/pullrequest/278112
2. Revert and update the [setskuimageversion vso util](https://devdiv.visualstudio.com/OnlineServices/_git/vsclk-core/pullrequest/280867) so that we could add keys in old format to the global database. This would allows us to continue using old key format with global configuration database.