Merged PR 281897: chore: installing auditd for investigations.
Fixes: [link to issue]

Description of the Change
-------------------------

chore: installing auditd for investigations.
Installing auditd in the pipeline. We can use that to investigate resume failures where process are getting killed.

Impact
------

10% of the resume are failing. It is not a hotfix candidate ATM, but is used for investigations.

Rollback Plan
-------------

This change it in the pipeline of linux vm creation. If you do revert this change, make sure there are not usages of it downstream.