Merged PR 278112: Appsetting to Use global backend database for configurations - still disabled
Fixes: https://github.com/microsoft/vssaas-planning/issues/1093

Description of the Change
-------------------------

This change allows the migration to be set via app setting which is disabled by default i.e. continue to use regional db
Impact
------

[How critical is the fix, think potential hotfix justification]

This change should not affect anything.
-------------

[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]

The changes can be rolled back by just reverting the commit or by adding a backend appsetting value as follows:

```
"backEnd": {
          "systemConfigurationMigrationSettings": {
            "useGlobalConfigurationCollection": false
          }
        }

```