Merged PR 281906: chore: always enabling process kill auditing for investigations.
Fixes: [link to issue]

Description of the Change
-------------------------

chore: always enabling process kill auditing for investigations.

Configure auditd for monitoring of killed processes
(rules don't persistent on reboot)
Since this is tracing syscalls, it takes CPU cycles. However, we don't expect kill syscall to be executed
in the order of 100s every second - so making this default for now. Perf of auditd is well studied and
based on the literature on the internet we shouldn't hit more that 0.1% CPU for this.

Impact
------
This is not a hotfix ATM. We are experiencing 10% resume failures. This will help with investigating the failures.

Rollback Plan
-------------
This can be safely rolled back. However that will limit us from investigating resume failures.