Merged PR 281993: [Bug Fix] Fix frontend archive continuation job handler state update error
Fixes: https://github.com/microsoft/vssaas-planning/issues/1026

Description of the Change
-------------------------

The original bug is:

In ArchiveEnvironmentContinuationJobHandler, we ignore the json serialization of ArchiveStatus in the payload. When "Continue", the ArchiveStatus will always be 0 initially. In RunCheckArchiveStatus, if the backend archive is still in progress, the ArchiveStatus will not be updated. We follow the pattern of setting payload next state from ArchiveStatus, so the next state of CheckStartStorageBlob will go to AllocateStorageBlob(0), that will cause creating lots of extra blobs and flooding the batch pool.

Fix:
Initialize ArchiveStatus from the payload current state.

Impact
------

Can fix the issue https://github.com/microsoft/vssaas-planning/issues/1026

Rollback Plan
-------------

Can revert