Merged PR 282439: Disable 1TB SKU on local dev stamp
Fixes: [link to issue]

Description of the Change
-------------------------

Not sure if related, but given we're at 100% capacity for compute resources in
WestEurope, and given that enabling this means that we need to create a new
pool/VM for any started dev stamp, better to disable it locally.

Impact
------

Very low, only affects local dev stamp

Rollback Plan
-------------

Yes, can be reverted.