Merged PR 282492: [Resources] Make CapMan include quota in lookups to avoid missing capacity entries blocking whole subscriptions
Fixes: [link to issue]

Description of the Change
-------------------------

Currently the capacity manager looks up capacity for all quotas within a ServiceType and then filters down to just the quota it wants. However, AzureSubscriptionCapacityProvider will throw if there isn't a record for a specific quota. This caused exceptions in Dev where a new quota was added to just one subscription - the other subscriptions threw not found exceptions even when the capacity manager didn't actually care about that missing quota.  The effect of this was that whole subscriptions were ignored in the capacity checks and capacity manager could only allocate in that single subscription.

Impact
------

Hotfix not needed, but it does need to be included when the new GPU SKUs are enabled

Rollback Plan
-------------

Yes, but see the Impact notes

Make CapMan include quota in lookups