Merged PR 282559: Disable lease in deprecated resource tasks with feature flag default value true
Description of the Change
-------------------------

When obtaining leases In the deprecated resource tasks, we call IsFeatureFlagEnabledAsync with feature flag default value true. If feature flag entry is empty or with value true in DB, we will not obtain lease for deprecated resource tasks.

Impact
------

We can eventually remove the flags from DB

Rollback Plan
-------------

Can revert.