Merged PR 282616: Enable polling for Queued state.
Fixes: https://github.com/microsoft/vssaas-planning/issues/977

Description of the Change
-------------------------
Enable polling for Queued state.
[Provide as many details as possible about what the code is doing, think about folks unfamiliar with the codebase]

Impact
------
Wihout this fix, on windows resume splash stays stuck "Getting things ready ...". It does not switch to connecting to the codespace once the codespace is available. It connects on refreshing.

[How critical is the fix, think potential hotfix justification]

Rollback Plan
-------------
Changes can be reverted.
[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]

Enable polling for Queued state.