Merged PR 281797: Configuration Writer commands - interactive and non-interactive
Fixes: https://github.com/microsoft/vssaas-planning/issues/1058

Description of the Change

Added a vsoutil that works in both interactive and non-interactive mode. In non-interactive mode, it will ask for any missing information. The interactive mode asks for information as and when required.

Impact
Should not affect anything in prod since it is just a util meant to be run by developers.

Rollback Plan
Can be reverted
[Can we simply revert this commit? If no please detail how to rollback. ⚠️ CRITICAL for DRIs]