Merged PR 282939: Use latest VSCode changes to use CDN urls instead of data: for iframe
Fixes: https://github.com/microsoft/vssaas-planning/issues/950

Description of the Change
-------------------------

We disabled iframe in !282427 PR to workaround the issues due to that with CSP we were seeing.
This was fixed in the latest VSCode changes to use IFrame from VSCode CDN instead https://github.com/microsoft/vscode-internalbacklog/issues/1529
This PR is to consume the same from our end.

Impact
------

This PR enables IFrame for web worker extension host which increases security (VSCode extensions will get limited access as they are in IFrame)

Rollback Plan
-------------

Revert