Merged PR 283000: [Web] Remove client validation of Cascade tokens
Fixes: https://github.com/microsoft/vssaas-planning/issues/1131

Description of the Change
-------------------------

We have client validation of tokens in our code which validate if the token has expired. And if the client time is incorrect then this validation will result in a valid token treating as invalid or otherwise. With this change we use the token to use API calls and when we get 401 we invalidate this token.

Impact
------

For customers with incorrect client time, they should be able to use our product without errors.

Rollback Plan
-------------

Revert