Merged PR 283337: Migrate to the new key names for native auth
Fixes: https://github.com/github/codespaces/issues/1467

Description of the Change
-------------------------

This change migrate to new key names that VSCode moved to for native authentication.

Impact
------

If this change is not present and the initial token that VSCode stores in new key expires the user will get auth failures even after suspend resume of the codespace.

Rollback Plan
-------------

Revert the change