Merged PR 283544: Kick off Environment Monitor Earlier in Suspend Action
Fixes: https://github.com/github/codespaces/issues/1490

Description of the Change
-------------------------

The state transition monitor will not repair states that have gone through "unexpected" state transitions. It's only looking for timeouts in expected transitions. We recently expanded the heartbeat monitor to catch environments stuck in the "Unavailable" state, and based on the the success of that change, we would like to expand the number of states that will be corrected this way.

Impact
------

Medium, corrects environments that are stuck in unexpected transitional states.

Rollback Plan
-------------

Rollback or disable feature flag.