Merged PR 283715: Avoid NRE when adding tags for snapshot records
Fixes: [link to issue]

Description of the Change
-------------------------

Recently added code to tag snapshots created from disks and disks created from
snapshots, for orphan tracking. This calls the patched method, and so was NRE'ing
because snapshots don't have pool details, so patch this extension method to not
NRE in that case.

Impact
------

Impact of the fix is low.

Rollback Plan
-------------

Yes, this can be reverted, although that would bring the NRE back.