Merged PR 283487: [web] Adding controller for serving static webview content
Fixes: https://github.com/microsoft/vssaas-planning/issues/390

Description of the Change
-------------------------
This change adds a WebviewController meant for serving static files in the webview iframe.
There will be a additional change needed to make use of this controller, that part is still being debugged.

This change also adds ingress rules for routing webview.{env}.codespaces.githubusercontent.com to the Portal service running in the cluster.

Impact
------

This change only does not affect the portal in any way. This simply adds a new controller but it will not be called from vscode yet.

Rollback Plan
-------------

Revert change and deploy.